分块上传接口Complete增加复用header逻辑；Copy&Complete接口增加校验逻辑 (#59)
* Copy&Complete接口增加校验逻辑

* 分块上传接口complete增加复用header逻辑

---------

+++ cos_c_sdk/cos_sys_util.h
@@ -68,7 +68,7 @@ int cos_url_decode(const char *in, char *out);
  * Ignores `locale' stuff.  Assumes that the upper and lower case
  * alphabets and digits are each contiguous.
  */
-long long cos_strtoll(const char *nptr, char **endptr, int base);
+int64_t cos_strtoll(const char *nptr, char **endptr, int base);
 
 /*
  * @brief Convert a string to int64_t.
@@ -81,7 +81,7 @@ int64_t cos_atoi64(const char *nptr);
  * Ignores `locale' stuff.  Assumes that the upper and lower case
  * alphabets and digits are each contiguous.
 **/
-unsigned long long cos_strtoull(const char *nptr, char **endptr, int base);
+uint64_t cos_strtoull(const char *nptr, char **endptr, int base);
 
 /*
  * @brief Convert a string to uint64_t.
